All service code has been moved to according service file
